small refactors and prep for bigger changes
