+++ api/settings/urls.py
@@ -9,7 +9,7 @@ from django.urls import include, path
 from api import views
 # Wire up our API using automatic URL routing.
 # Additionally, we include login URLs for the browsable API.
-from api.views import UserListView
+from api.views import UserListView, UserApiView
 
 urlpatterns = [
     path('api/auth/', include('rest_framework.urls', namespace='rest_framework')),
@@ -17,5 +17,6 @@ urlpatterns = [
     path('api/students/current/', views.CurrentStudentView.as_view()),
     path('api/students/current/hours/', views.CurrentStudentHoursView.as_view()),
     path('api/mentors/', views.MentorListView.as_view()),
-    path('api/users/', UserListView.as_view(), name='user_details')
+    path('api/users/', UserListView.as_view()),
+    path('api/details/', UserApiView.as_view())
 ]
