+++ auth_backend/modules/user/views.py
@@ -3,17 +3,12 @@ File Name: Views
 Purpose: Django views for rendering a variety of data.
 Comments:
 """
-
 from django.urls import reverse
 from django.views.generic import TemplateView
 from django.views.generic.edit import FormView
 
 from auth_backend.modules.common.mixins import LoginRequiredMixin
 from auth_backend.modules.user.forms import ProfileForm, StudentForm, MentorForm
-from rest_framework.views import APIView
-from rest_framework import permissions
-from rest_framework.response import Response
-from .serializers import UserSerializer
 
 
 class ProfileCreateView(FormView):
@@ -75,14 +70,3 @@ class ProfileMentorInfoView(LoginRequiredMixin, FormView):
         mentor.save()
 
         return super().form_valid(form)
-
-
-class UserApiView(APIView):
-    """
-    API endpoint to retrieve user info
-    """
-    permission_classes = (permissions.IsAuthenticated,)
-
-    def get(self, request):
-        serializer = UserSerializer(request.user)
-        return Response(serializer.data)
