+++ frontend/src/components/elements/Cards/userPic.js
@@ -4,10 +4,11 @@ import Container from 'react-bootstrap/Container';
 
 class UserPic extends React.Component {
     render() {
+        const imgSrc = this.props.imgSrc
         return (
             <Container>
                 <Image width={275} padding={0}
-                       src="http://winkeyecare.com/wp-content/uploads/2013/03/Empty-Profile-Picture-450x450.jpg"
+                       src={imgSrc ? imgSrc : "https://www.insidemathematics.org/sites/default/files/2019-10/610-6104451_image-placeholder-png-user-profile-placeholder-image-png.png"}
                        roundedCircle/>
             </Container>
         )
